refactor: extract pane content portal logic into usePaneContentPortals hook
